Optimize ClientBuilder
